requirements.txt
numpy~=1.24.2
pandas~=1.5.3
matplotlib~=3.7.0
scipy~=1.10.1
requests~=2.28.2
bs4~=0.0.1
beautifulsoup4~=4.11.2
yt-dlp~=2023.2.17
youtube-search-python~=1.6.6
pytube~=12.1.2
yolov5~=7.0.9
python-dotenv~=1.0.0
ydb~=3.1.1
multi-rake~=0.0.2
scikit-image~=0.20.0